Never too early to refactor a bit
