--- Utils/StaticDevice.cs
@@ -1,8 +1,0 @@
-using LoupixDeck.LoupedeckDevice.Device;
-
-namespace LoupixDeck.Utils;
-
-public static class StaticDevice
-{
-    public static LoupedeckLiveSDevice Device { get; set; }
-}
